requirements.txt
-i https://pkgs.safetycli.com/repository/la-zarza/project/etl-sad/pypi/simple/
# Excel ETL Dependencies
# Core data processing
pandas>=2.0.0
numpy>=1.26.0

# Excel file reading
openpyxl>=3.1.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration management
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.2

# Email notifications (built-in Python libraries - no additional install needed)
# smtplib, email.mime
